Move KRLanguage class to a different package
