+++ src/main/java/nl/tudelft/goalkeeper/parser/results/parts/Expression.java
@@ -3,7 +3,6 @@ package nl.tudelft.goalkeeper.parser.results.parts;
 import lombok.Getter;
 import lombok.Setter;
 import nl.tudelft.goalkeeper.checking.violations.source.Source;
-import nl.tudelft.goalkeeper.parser.queries.KRLanguage;
 
 /**
  * Abstract class for expressions.
