+++ src/main/java/nl/tudelft/goalkeeper/parser/results/parts/Linguistic.java
@@ -1,4 +1,4 @@
-package nl.tudelft.goalkeeper.parser.queries;
+package nl.tudelft.goalkeeper.parser.results.parts;
 
 /**
  * Interface for classes having a KRLanguage.
